AssignID refine and Dot90
Refined AssignAllIDs() so that if the first row of POI's is missing, the order is still correct (now based on z position, not position in list of POIGroups). Also changed dot product attention to angle mapped to 1 to 0 for linear measurement to centre (dot product is cosine measure to centre).